move DID base class into PLC dir
